graph_desc_parser_parse written to end
Not yet tested.
Still need to see about parsing arguments to nodes.
In addition, it might be better that the copying be even tighter.
The way we'd do that is determine the length of all the node names (and
eventually also the nodes' arguments) and allocate enough space to hold these
strings. Then point each node name to the starts of these strings.
It's probably okay to have a memory-heavy parser, but it might be better for the
many graph descriptions that evenutally will be floating around in memory to be
have smaller memory requirements.