Wrote a bunch of shit having to do with parsing graph arguments
It doesn't yet work. This may never work exactly as described here but some
parts are useful for future work. In particular the "string compression" which
moves variable names to later memory so that they can serve as keys in a
variable table. The string is then shortened and where the variable expansion
should take place in the string is stored.